requirements.txt
boto3
requests
python-dotenv
Flask
Flask-Cors
streamlit
requests
pandas